requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
pyaml
# -e .